SINQ: Use more code in simulation/dry run
If there is no direct hardware access the 'hardware_access = False'
should be set to run the 'do...' methods in the classes. These methods
will be called in the simulation/dry run. This allows better tests and
dry run results.

The method calls of the attached devices shouldn't got through the
'do...' methods, which may lead to a non safe call in simulation.

Change-Id: I1a3c2aebe1824a8ba8875f2c22c8aba7cce22672
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/32126
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>