TESTS/ESS: Fix not raising error in log handler test
Newer versions of the 'kafka-logging-handler' package (> 0.2.4) do not
raise the expected exception in case of not existing kafka host or
connection.

Since the 'kafka-logging-handler' package doesn't have a version info we
have to check the used version via the pkg_resources package which asks
the python package database about it.

In the requirements is a limitation for the package version of the
'kafka-logging-handler' and for the most cases the check seems to be
needless but we are prepared for the future.

Change-Id: I436289cb60f24e5153a3bba97857a9d1c38ddd3d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/32333
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>