new magnet base class
Needs a magnetic probe.
This magnet class supports bipolar power supplies and uses
experimentally obtained Field(Current) curves for operation.
The calibration procedure takes several minutes of time, and in
general doesn't have to be repeated without relevant change in
the equipment. The dependancies can be measured at different
current ramps and are stored in NICOS cache for convenience.

Change-Id: Idbfee77f86990890b42639cfd9573a1c45213b3d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/32690
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>