Console: Add feature to add a selected part to the elog
There was a very heavy wish ( ;-) ) to add parts of the console output
which can be selected to attach to the electronic log.

The background: For some implementation of the elog backend the normal
messages will be suppressed, but for some reason the users want to add
some output in the console to their elog.

Change-Id: Ie0d1ae500862b1af55047d5f6b6ea56a3e35e410
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/32777
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>