Test: Make the ILL sink test locale independent
Since the test checks for some date entries containing the name of the
month which is written in English the environment has to be set as well
to English and UTF-8. If this is not the case the test may fail.
The spelling for December is a little different in English and German
for example.

Change-Id: I8e8340fa47f32f449a4449eb0d06b7b06d2f43a2
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33116
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>