HostnameMailer: Make the 'subject' parameter handling more NICOS like
The parameter 'subject' in HostnameMailer should be volatile to get the
right value (at the moment: hostname) at anytime.

The previous code had the disadvantage, that any change of the content
of 'subject' did not have any effect on the '_subject' member. At the
moment the 'subject' parameter is not settable but who knows about the
future or derived Mailer classes.

Change-Id: I1bb08e7a5414793387ce5d6c9e64ab221e4cd020
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33188
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>