Gap: Add parameter 'min_opening'
For some of the gap devices it is allowed to "overlap" the blades. Since
this is allowed by design the possibility should be given. If
'min_opening' is negative, this will be not any longer forbidden for the
device. In case of a gap which can't be really closed, the 'min_opening'
is a number > 0. The current state is given by the 'min_opening' set to
0, which is the default.

Change-Id: Ifebf9ef8e974ce8d8f603318ae163f69b2cb4a5a
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33305
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>