bypass qt bug in QMainWindow in macos
This change demaximizes the QMainWindow before settings are saved on
macos.

Related to https://bugreports.qt.io/browse/QTBUG-4397
Despite the bug is mentioned as resolved, on macos, when an instrument
configuration is chosen, if a user exits from nicos-gui while nicos-gui
was in maximized window mode, something wrong happens to saveGeometry(),
and later the geometry can't be loaded again, which crashes the
nicos-gui, when user tries to select a GUI configuration related to the
instrument on next start.

Change-Id: I6700e5552a9d74aa9c45d44ae6ddabfffe1fd650
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33377
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>