pylint: Remove the comment for unnecessary lambdas
There are a lot of places where this message(s) pop(s) up in case there
was a previous overriding of this message(s) by one of the reviewers.

It seems, that these messages not in all cases are quiet helpful and an
ovrriding at each of theses places can be very expensive in time.

Change-Id: I3991f8f32c83d8aaddea5014cf23b10ca06759ed
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33496
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>