moke01: baseline feature
This has new tab in QTabWidget implemented which invoked little
rfc of the ui logic.
Feature adds measurement and handling of baseline curves.
Baseline curve is required because instrument consists of many
metal parts, which produce constant offset on measurements.
The baseline curve would be different for different measurement
modes, likewise the calibration curve. And therefore is handled
similarly.
To measure the baseline, a standard measurement should be
performed. Then to store new baseline, the last measurement
should be imported in a proper ui section, and then saved.
Current baseline is stored in a cached parameter of a MagB nicos
device. At every measurement the proper baseline is written to
the MagB measurement cached parameter. This allows to subtract
the baseline at any moment from the raw data and use resulting
curves for subsequent analysis or export.

Change-Id: I9a9375ac95f4717ff9b87189b53330fd321a976f
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33655
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>