moke01: code rfc
Now when MokeMagnet performs the measurement, measurement mode is
stored in the class parameter as it should.
In `stepwise` measurement mode `_hw_wait()` is invoked on every
measurement point and the first measurement point is not set
twice anymore.

MokePanel doesn't need `read_last_measurement()` method, because it
shares same functionality with `_update_intensity()` method. One
method `_update_measurement()` should suffice.

`MagB._IntvB` is not a cached parameter, when the daemon is restarted
this value is []. Added extra logic to handle that.

Change-Id: I36640f58f58c406010824f1d9a95fb84b4df9f4c
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33663
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>