GR: Y axis label is not covered by labels from Y axis ticks
fixes: #2428
The patch adjusts the viewport and offset of Y label depending on
tick labels lengths.
Tick lengths are placed to buffer, which segregates current ticks
from the past ones. The longest tick label from the buffer is
used to calculate offset of Y label and viewport.

Needed to increase viewport ymin to 0.18 otherwise pygr.Plot.drawGR()
goes in RecursionError.

Change-Id: I7547068671d3611991f099d1b16874a5e6cb9080
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35408
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>