MLZ: Add special setups
Since there is a very the case that the hardware will be switched off
during a instrument downtime, but the services are running as before
there will be produced a lot of error messages it would be helpful to
give the users (instrument scientists, technicians) some special setups,
that may be loaded for these cases:

- shutdown.py: The instrument will be shutdown and in priniciple the
  intrument is "dead"
- standby.py: The instrument will be set into a state where only a
  minimal set of devices is needed (pumps, sensors, ... ) which are
  needed.

At the moment the setups are empty, but belonging to the 'basic' setups.
They can be filled up with the need of the specific instruments.

Change-Id: I0bc621acb258601ce822611113c4c7be4f6771d3
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/33762
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>