modify handling of errors on reading parameters
ReadFailedError is no longer taken into account for the warning
in the device status. Errors when reading a parameter might happen
in a foreseeable way, e.g. reading a setpoint when the control is
not active.
In these cases, this should not be taken into account for the
warning about errors on parameters in the status, because it
might worry the user.

Change-Id: Iac50bb30f695de2a2a385db25db0ff0ba881ddd1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34020
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Markus Zolliker <markus.zolliker@psi.ch>