DataSinks/Image: Clean up '_processArrayInfo' code
Since the call of the _processArrayInfo method is always done with the
'sink.detector.arrayInfo()' as parameter there is no need to do the call
inside the method again.

The check for the length of the arrayInfo has been changed for the check
to length equal 1 to avoid problems with the access to the first array
in case the length is 0.

Change-Id: I67d6eb9ec63c91cdaa93b7c01cbc7803453d6fcb
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34066
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@tum.de>