Virt/Reseda: Refactor image data reading
The file writing is not the same as in standard McStas, so the first
data will be written only on SIGUSR2 signal. The normal McStas reading
writes a file immediate after start (but more or less empty).

The warning about the missing bin file should be only given if a SIGUSR2
was sent but no file found.

Change-Id: I246b35e64bc9531bd7b26991ecf2cc3b002544de
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34140
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@tum.de>