SECoP: Fix explicit device creation
Explicitly defined devices did not get the configuration on the SEC
node, with parameters missing validators etc. afterwards.

Device config from the setup file is now merged into the module config
from SECoP, so that manual and automatic devices do not differ.

This change removes the ability to not expose a SECoP parameter, this
has to be re-introduced if needed.

Change-Id: Iefcb74ee7cbbb1c06d534128533d6e0e708cf74d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34272
Reviewed-by: Alexander Zaft <a.zaft@fz-juelich.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>