NEXUS: bug fixes and better protection against errors while writing files
Sometimes, NeXus file writing runs into an error and then does not close
the file properly. Which can only be resolved by restarting the daemon.
Moreover, the desired behavious is to write as much of the file as possible,
even in the presence of errors. This patch attempts to implement this.

Change-Id: I9444939b69c010e5f9a9dde75c954bebf8508c9c
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34512
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@tum.de>
Reviewed-by: Mark Koennecke <mark.koennecke@psi.ch>