test reruns might still fail when `only_rerun` condition is used
If used as `only_rerun='AssertionError'`, the condition seems to be wrong
and the test fails normally. Apparently AssertionError must be passed as
an object, despite what is given as an example at
https://pypi.org/project/pytest-rerunfailures/.
I suggest to not use this condition, because I cannot verify its sability.
When I run locally the pytest with `--count=10000`, when `only_rerun`
condition is specified as an object or a list with one object, the pytest
itself fails eventually, at least on my setup. Yet simple
`@pytest.mark.flaky(reruns=5)` works fine.

Change-Id: Ib43cb1a0a78a4e7fdbb44d347e8d6ca1c994cc47
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34562
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>