LImA: Automatic offset detection in image read
LImA transfers the offset to the image data in its data stream getting
from the readImage function. In version 4 the header size was changed so
the current header size (fixed coded) led to a problem during the
conversion to a numpy array.

Now the code should be able to handle the old versions as well
correctly.

Change-Id: I13a998999a6411907c801be9f28e7283fe21d810
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34627
Reviewed-by: Jens Krueger <jens.krueger@tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>