Test: Improve data fit test
The test to fit a peak from scan data is a little flaky due to the
randomized counts during the cscan.

The stepsize as decreased but the number of steps has increased, so we
have the same scanning range but more data points. This may help to
reduces the 'flakyness' of the test.

The runtime of the test increases a little from ~2 secs to ~3 secs per
test run. If there is no rerun needed anymore, in total the test may now
a little faster.

Change-Id: I2c5705b2fe909949b3dbb642d675a30c4ec3e642
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34656
Reviewed-by: Jens Krueger <jens.krueger@tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>