Physlab: Add special shutter device implementation
The shutter should only be opened if both doors are closed. If this is
not the case the HV controller will run into trouble. It's not a safety
issue, but the controller will go into a state where a hardware restart
is needed and the whole system has to be realigned :(

Change-Id: Ice5918b745947cf9cce89a0ec3425a1394ba7ba9
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34672
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@tum.de>