Physlab: Read hardware register on every door read
The doors of the X-ray machines are very important for the radiation
safety and should be closed before opening the shutter. The shutter
itself can't be opened if one of the doors is not closed properly. There
is a hardware interlock, but the HV generator may run into a state,
where it can't be resetted only by power cycling, which requires a full
new instrument alignment. So it's very important to get the right
information about the door state. Every read of the state should ask the
hardware registers.

Change-Id: Ie7bbb06890d5b9b5aec32115010d582bba248cd5
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34673
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@tum.de>