DOC: Change the order of methods in the GUI client
If the someone reads the doc of the GUI client API he will be guided
first to the lowest level methods the get some information from the
server. In most cases it would be much better and simpler to use the
provided higher level methods, but the reader won't go down to the
higher level methods, if he/she finds the lower level methods sufficient
to solve the problem.

Change-Id: I73bae6c11b9d7db5d7a3f20a3a95266b1cc1a118
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34795
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>