session: fix recursive calls when cache connection closed
When the cache is disconnected, it logs a message, which is sent
to the ELogHandler, which calls elogEvent.  But checking
`self.experiment.elog` makes a call to the cache.get() which again
will complain with a log message when there is no cache connection.

This leads to NICOS not shutting down cleanly after the cache is
already dead.

Change-Id: Ic5943b298c7cc87f5c89c34c1bf2ec3768b5e6b9
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/34899
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>