SINQ: Primarily updates to the Eiger instrument configuration files from the last few months.
In `nicos_sinq/devices/illasciisink.py` I wondered if we should at least add
some warnings when a device isn't found when logging scan results, as it seems
that for quite some time, files were being incorrectly written.

The code in `nicos_sinq/eiger/commands.py` likely doesn't work.  It was added
by the instrument responsible, and I don't want it to be lost in the next
release. It is unlikely that there will be any further work on this until next
year when I have a chance to sit down with them.

Change-Id: Ia5f7df46b16ed2c802f040db20c5b79b0259bbd4
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35047
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>