TAS Commands: Extend HKL parameter type set/checkalign
This makes the command a little bit more flexible in HKL parameter type.

Due to the use of the 'Q' command the parameter type of the hkl parameter
can be more or less the same as the for the 'Q'. The 'E' value will be
set in any case to 0.

Change-Id: I1113fce6b7778473d93c0caa778758c7420262e6
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35058
Reviewed-by: Jens Krueger <jens.krueger@tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>