EPICS: Added polling for motor and speed limits
In the EPICS motor record, both position and speed limits are volatile.
To account for that, a poller has been added which regularely updates
these limits.

Additionally, the speed limits are now visible to the user
and the error message in case an invalid speed has been set has been
improved.

Lastly, graypy is not needed to run test_epics_motor.py, so the pytest
skip clause has been removed.

Change-Id: I2380f5bd48a842c4f12c1fa6a083129fbd20fe96
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35260
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>