SINQ: Rebuild the SINQ version of EpicsMotor (new driver architecture)
The Epics motor driver at SINQ has been revisioned and therefore the
Nicos equivalent needs to be adjusted as well. In particular, the
following changes and bugfixes are implemented:
- Starting Nicos does not autoenable the motors anymore. This is a
request of the instrument scientists
- Some motors cannot be disabled. This information was previously stored
in the Nicos motor configuration, but is now stored on the driver level
and Nicos simply reads this information from the driver.
- The encoder type is now read from Epics and is not needed in the Nicos
motor configuration anymore.
- The intermediate state "disabling" / "enabling" has been removed. When
restarting the IOC, the "Enable" PV could be 0, but the "EnableRBV" PV
could be 1, which lead to the motor being shown as permanently disabling
the motor. This could only be resolved by actually disabling and then
reenabling the motor. Now the motor simply shows whether it is enabled
or not (by just reading EnableRBV)
- Motors must not be disabled during movement. This is already prevented
on the Epics driver side, but now Nicos checks this as well and returns
a warning message to the user (previously disabling during movement
silently failed).

The old version of EpicsMotor is still available as
nicos_sinq/devices/epics/motor_deprecated.py, since the new SinqMotor
depends on new Epics driver versions which are not available for all
motor types yet. The setup files have been changed accordingly to use
the deprecated version for now.

Change-Id: I5ff5465f555dc2d9d96082ba6f1a58aa6ab1f47f
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35285
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>