MLZ/NeXus: Create base NeXus structure
There are some components which will be used in different NeXus
structures.

The main basic structure for the MLZ NeXus contains the reactor source
and uses some information about the instrument and user structure which
are used at all MLZ instruments.

For some instruments the NeXus templates are very similar, only the
devices names are a little different. For these common NeXus template
providers are created.

Change-Id: I3e503e5939b1fdd22f57fa6ea64d137e4c71e359
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35699
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>