GUI: Don't resort values for the display
The resorting of the values for 'valuetype' in 'MappedMoveable' and in
the ComboWidget lead to unexpected behavior. The dictionary in Python
now (since 3.7) is an ordered one, so it would be easier to predict the
behavior, if the dictionary entries (in code or configuration) are
ordered in the requested order.

Fixes: #4963
Change-Id: I596007fe8b205811f70c5941bda8b9e5b9d37da5
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35657
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>