moke01/output: changes
Fixed bug which was hiding new values during measurement.
Min and max intensity are not needed.
Canting angle is output in SKT and mrad.
There are 2 possible values for number of cycles: initially
planned number by the user and number of completed cycles.
Sometimes user can abort the measurement, and this change
serves to indicate it.
`Sample id` not `id`.
Measurement names starts with year, 2 digits are enough.
Values `Bmin` and `Bmax` do not need extra explanation.

Change-Id: I46ea54372dcb8ac8479127c3b5177ff5d668d79a
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35686
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>