GUI: Remove the 'facility' option from configuration
The facility option was used to indicate to use a FlowUI specific QSS
file.

The change checks the type of the 'MainWindow' class and if it is
derived from the FlowUI MainWindow class it uses the QSS file from the
FlowUI file in any case. The reduces the configuration effort, since the
'facility' option was set to all SINQ and ESS instruments (using the
FlowUI).

Change-Id: Ibbfd40662233a359525ada34150027730d11f265
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35705
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>