CORE: Added mapping to doStatus of Switcher class
The switcher class does map the value of the underlying device for the
"Current value" in the GUI, but not for the status message if e.g. the
device is moving. For example, if "0" is mapped to "off" and "1" is
mapped to on and the underlying device reports "moving to 1" when
switching from "off" to "on", the current version of the Switcher class
reports this same message. This PR tries to deliver a better user
experience by replacing the "1" in the status message with "on" (so that
the switcher device reports "moving to on").

Change-Id: I28610153a1bebd96cc4f2bb799d2fb9cee458530
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35831
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>