auth: fix error msg with multiple authenticators
Having two authenticators configured overwrites the errors of the first
one, even if the login was successful and other factors prevent the
login. Therefore still the message will be "invalid username or password".

Example would be Ghost and base Authenticator: Successful ghost login
but without a valid proposal gives the message from the base auth
afterwards.

So, distinguish between failed login and post-auth errors with a new
subclass.

Change-Id: I37009edb8575886a3ec535be2548c48f6175f894
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35938
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Alexander Zaft <a.zaft@fz-juelich.de>