CORE: Cover EpicsMotor edge case maxspeed==0
maxspeed == 0 in the EPICS motor record means that the maximum velocity
range checking is disabled. In order to conform with this definition in
NICOS, we check if maxspeed from the motor record is set to zero.
If that is the case, we set maxspeed to infinity, because maxspeed == 0
means that there is no upper speed limit.

Change-Id: I94a2359a5d3eef608a4e26b04c6b59b988cce73d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35942
Tested-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Muhammad Septyawan Aulia <muha206@brin.go.id>
Tested-by: Muhammad Septyawan Aulia <muha206@brin.go.id>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>