SINQ: Read precision from MotorRecord
Since the motor record itself already provides a precision in form
of the MRES field, this PR overwrites any cached precision value
with the current MRES field. This avoids the "Did not reach target"
warning in case target and actual position do not match exactly and
the cached precision is too small (e.g. if the cache has been deleted,
the precision is initialized to 0!).

Change-Id: I1ed359ec2bf251adfbc62748f9cb22958c5c769e
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/35968
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>