SINQ: Fixed calling method of non-parent class
`PitchButton` used to call the `setClient` method of `ValueLabel`, even
though the latter is not a parent class of the former. Since the
`setClient` method of `ValueLabel` just calls the one from `NicosWidget`,
changing this does not change the behaviour of `PitchButton`, but makes
more sense from the perspective of the inheritance chain and guards
against later implementations of `setClient` in `ValueLabel`.

Change-Id: I7587b00ae90be4227ae0892b1d17eda842ca9706
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36014
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>