Adjusted docstring for class Param
In general, we want to use triple quotes for docstrings instead of
triple primes. The docstring example for `ext_desc` of the Param class however shows the use of triple primes
(likely due to collision with the triple quotes of the docstring
itself). This patch uses escape characters in order to display the
`ext_desc` example with the preferred triple quotes.

Change-Id: I06f9f610b60ea61a63295abe42e99996c737f534
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36090
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>