SINQ: Fixed illasciisink for count command
The illasciisink device didn't work together with the count command because it was expected that the field session._currentscan was set (which only happens in the scan command, not in the count command). This patch therefore checks whether we are doing a scan and adjusts the header accordingly.

Also, some small setup fixes in TASP were necessary.

Change-Id: I67fafee60c427646fdb728f39c6f145d4a2233f1
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36194
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>