CORE / SINQ: Read precision of EpicsMotor from motor record
The precision of an EpicsMotor is defined within the motor record field
MRES. Therefore it should be read from this field as well and not be
writeable, since changing the precision at runtime does not adjust the
limits correspondingly (see motor record docs:
https://epics.anl.gov/bcda/synApps/motor/motorRecord.html#Fields_res).

In case this patch runs the danger of creating bugs at other facilities,
I can also move the entire code to SinqMotor. However, since MRES is not
a SinqMotor specific field, I wanted to generalize it first.

Change-Id: I55ae4dc3f34ca4588dcc5a36e8489a23e639e6e3
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36267
Tested-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>