CORE: Give EpicsMotor errors priority over "starting" msg
If starting an axis fails, the previous setup does not show the error
message immediately, but only that the axis is "starting". Only after
the starting delay is over, the error message is shown. I propose that
the error message is shown immediately. At SINQ, we have some
non-persistent error messages which only "flash up" for a short period
of time and I don't want them to be masked by the "starting" message.

Change-Id: Ic352f7b0c1f7deb41500538a9d184306b0034ed5
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36359
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>