Test: Add test suite to verify code for wrong 'overrides' detection
The code to detect the invalid use of parameter overriding has to be
verified to detect the incorrect use.

The override of a parameter is only allowed in cases the parameter
really exists in one of the inherited classes and/or mixins.

The override can be defined in a class but in a used mixin as well.

Change-Id: Id806f358c00bd3d61ce928ce01ebb7c1acb5ced8
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36407
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>