SINQ: Added isConnected property to SinqMotor
Our new controller generation (which we also control via the SinqMotor
interface) allows for motors being disconnected during operation (e.g.
while NICOS & EPICS are running). The controller then also reports that
the axis is disconnected and we forward this information via the
"Connected" PV. Therefore, we also need to reflect this in NICOS, which
is done in this patch.

Additionally, the extension PV handling was reworked and is now based on
a dictionary instead of a long elif chain. Lastly, optional PVs
(errormsgpv, errorbitpv, reseterrorpv) now have dynamically generated
default values.

Change-Id: Ie9f38b8ee30d62034faa7b7d0dc7fc6d77908be2
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36249
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>