gui/help: fixes
No need to check if url is QUrl.

QWebEngineView doesn't have `scrollPosition` method, as it was
the case in QWebView in qt4. In qt5/6, QWebEnginePage has it.

QWebEnginePage class didn't have `mainFrame` since qt4, until
it appeared again in qt6.8. So the logic of `HelpWindow.showHelp`
started to fail, since `page().mainFrame()` started to return
QWebEngineView instead of QWebEnginePage, and the former doesn't
have `scrollPosition`.

It also looks like handling of url starting with `#` never happens
since QWebEngineView doesn't have `findFirstElement` nor
`setScrollPosition` methods, so we should be getting errors.
And seems we do not.

Change-Id: Ic7fa8ec3a8ed38eb7d29dd5fcebfab4ede4fbcba
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36573
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>