SINQ: remove alias_config for a3/om
Selecting the desired alias target in the setup panel does not work
properly with dynamically created devices (SECoP). It is therefore
better not to add a3/om to the alias_config. There is only one
static possible target, the other one (or possibly two) are
dynamically created. Changing alias targets for these have to
be done on the command line.

Change-Id: I34450fb39cbbea4b17798bd3f77306899cea55af
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36602
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Markus Zolliker <markus.zolliker@psi.ch>