SINQ: Reset parameter ccl_file on stop
The "Measure" command in nicos_sinq/sxtal/commands/__init__.py does set
the parameter ccl_file to True before starting a measurement and to
False after the measurement is done. However, in case the device is
stopped (e.g. by an emergency stop), the parameter is not reset to
False. This patch fixes that.

Change-Id: I082f6b9d571938c4f3398a81f0b5766784ce7bcd
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36642
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>
Reviewed-by: Edward Wall <edward.wall@psi.ch>