SINQ: Remove DynamicUserLimits from SinqMotor
The Mixin DynamicUserLimits is currently still bugged (see the offset
problem handled in https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36622)
and is only needed for a minority of applications. Even there, it is just a quality
of life feature and not strictly needed to operate the devices.

Therefore, this patch removes it entirely from the pyepics-based SinqMotor class
to avoid the offset bugs which are known to cause problems in the field
(i.e. setting an offset larger than the limits can lead to weird
behaviour).

Additionally, I also removed polling the userlimits, which was a
mistake, since it can result in the poller undoing offset-caused changes
of the userlimits (this was discussed in a Tuesday meeting and field
testing has found the polling to be a root cause of these problems).

The DynamicUserLimits mixin will then be included in the caproto-based SinqMotor class (see
https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36613).

Change-Id: I128cb1f13dd3470ac4354695d914a99ffe4bd15d
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37009
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Tested-by: Stefan Mathis <stefan.mathis@psi.ch>