CORE: Added deadband handling to doIsAtTarget of EpicsMotor
The old doIsAtTarget method of EpicsMotor only considered the motor
resolution, not the deadband (which might be way larger than the motor
resolution). This patch rectifies that by using the larger of the two
values in order to determine whether the motor is at target or not.

Under the hood, the motor record field SPDB is used:
https://epics.anl.gov/bcda/synApps/motor/motorRecord.html#Fields_motion

Change-Id: I3f05ba703516f45ddc369f293f12b9ed5502af47
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37368
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>