SINQ: Fixed pause bugs in EpicsDetector
When using "Pause while counting" or "stop right now", the "pausepv"
gets set to 1 (doPause is called). If one then uses the "Stop" button in
the top right corner, the "doStop" and "doFinish" methods get called.
This then leads to a situation where the device is idle from a NICOS POV but
still paused from the driver / device POV. This desync then means that
no new count can be started anymore.

This problem was encountered the first time after
https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36748
but is in fact completely independent from that change (it only
highlighted the issue).

This patch provides a quick fix by simply switching out of the pause
mode if doStop or doFinish is called. A proper solution would be
resolving the issue directly in the EPICS driver, however this driver is
going to be phased out anyway (replaced by
https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/36209)

Change-Id: I1538efbc847db0a39a20ce24ece9445353e62618
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37418
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Edward Wall <edward.wall@psi.ch>
Reviewed-by: Enrico Faulhaber <enrico.faulhaber@frm2.tum.de>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>